src/training/requirements.txt
pillow
google-cloud-storage
google-cloud-bigquery
pythonjsonlogger>=2.0.7


